django-rest/requirements.txt
djangorestframework==3.13.1
django-cors-headers==3.12.0
django-health-check==3.16.5
mysqlclient==2.1.0
Django==4.0.4
pytest==7.1.2
pyjwt==2.4.0
